move loader path
