+++ langchain_docugami/__init__.py
@@ -1,4 +1,4 @@
-from langchain_docugami.document_loaders import DocugamiLoader
+from langchain_docugami.document_loaders.docugami import DocugamiLoader
 
 __all__ = [
     "DocugamiLoader",
